Tested VR and enabled workbench/tools
Tested a VR build to ensure it worked, and once I did I reenabled the workbench and tools so they would appear in the scene instead of the default cube you add in the tutorial. Then I had to add some "building blocks" to the tools so that they were grabbable, and debugged the spray bottles so that they were affected by gravity once dropped. Required me to make them not kinematic.