Added screen progression
Created a prefab for the question object, so if something is changed like button size or text font they will all be affected. Also added a simple script for progressing screens, attached to the Background image object. add it to buttons to use the NavMesh scripts methods. add the current question to close, and / or a new question to open.