Finished the Wet Bench
The wet bench has been fully realized in VR with the hinged door and sliding doors with their own components found inside of them. The components are static.

Added materials to the wet bench such as a stainless steel look to all the metal components, color to the doors, the solvents tag, and the FLAMMABLE warning labels

What else needs to be completed:

 - Fully implementing the quiz questions and metric tracking