Even more fixed
I don't know why this wasn't the very first thing I checked, saving me from this entire 7 hour long mess. but the issue was from the listener component being turned off. that's it. oh my gosh. also I forgot to mention this in the previous commit, but the reason it wasn't letting me pull the trigger was because the static methanol props had a bunch of scripts on them that were for the actual bottle and so were broken. and... they were on there because I updated the prefab for the bottles and that was the last thing I pushed and I didn't bother playtesting after that and oh my gosh I know everything now. but again audio is fixed now.