video transision functionality added.
Added a script to allow for the video to play at the beginning, and then transision to a screen asking whether to replay the video or start this quiz. after your done with the quiz, pressing restart takes you to this screen. Also, perhaps foolishly, I moved the assets to their own folder.